# openapi/auth.yaml
openapi: 3.0.3
info:
  title: Insforge Authentication API
  version: 2.0.0
  description: Authentication endpoints with separated auth and profile tables
paths:
...
  /api/auth/oauth/google:
    get:
      summary: Initiate Google OAuth flow
      tags:
        - Client
      parameters:
        - name: redirect_uri
          in: query
          schema:
            type: string
            format: uri
          description: URL to redirect after authentication
      responses:
        '302':
          description: Redirect to Google OAuth
        '500':
          description: OAuth not configured
  /api/auth/oauth/github:
    get:
      summary: Initiate GitHub OAuth flow
      tags:
        - Client
      parameters:
        - name: redirect_uri
          in: query
          schema:
            type: string
            format: uri
          description: URL to redirect after authentication
      responses:
        '302':
          description: Redirect to GitHub OAuth
        '500':
          description: OAuth not configured
  /api/auth/oauth/discord:
    get:
      summary: Initiate Discord OAuth flow
      tags:
        - Client
      parameters:
        - name: redirect_uri
          in: query
          schema:
            type: string
            format: uri
          description: URL to redirect after authentication
      responses:
        '302':
          description: Redirect to Discord OAuth
        '500':
          description: OAuth not configured
  /api/auth/oauth/twitter:
    get:
      summary: Initiate Twitter (X) OAuth flow
      tags:
        - Client
      parameters:
        - name: redirect_uri
          in: query
          schema:
            type: string
            format: uri
          description: URL to redirect after authentication
      responses:
        '302':
          description: Redirect to Twitter OAuth
        '500':
          description: OAuth not configured
...
  /api/auth/oauth/{provider}/callback:
    get:
      summary: Provider-specific OAuth callback
      description: Alternative callback endpoint for specific providers
      tags:
        - Client
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [google, github, discord, twitter]
        - name: code
          in: query
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
        - name: token
          in: query
          schema:
            type: string
      responses:
        '302':
          description: Redirect with access token
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKey:
      type: apiKey
      in: header
      name: x-api-key
...